added a function to determine delta_t for a specific date The function uses a table of historic delta_t values ranging from the year -2000 to 2022 (data from IERS) Beyond this I use dt=-20+32*t^2 where t is the number of centuries since 1820 see L.V Morrison & F.R. Stephenson, 2004, Journal for the History     of Astronomy, 35, 327-336
Changed routines such that:
delta_t is a pointer. If the pointer if NULL freespa computes
a delta t value. If the pointer is not NULL it will delta t
equals the value found at the end of the pointer

Note of no practical relevance:
NREL spa limits delta t to +-8000. My best guess is that this
limits its use to the period from 300-3400 (or thereabout).